+++ bergson/scorer.py
@@ -1,135 +1,85 @@
-import json
-from pathlib import Path
-from typing import Callable, Literal
+from typing import Callable
 
 import torch
-from datasets import Dataset
 
-from .data import QueryConfig, load_gradient_dataset
-from .gradients import GradientProcessor
+from .data import QueryConfig
+from .score_writer import ScoreWriter
 
 
 class Scorer:
+    callback: Callable
+
     num_scores: int
 
-    def __init__(self, callback: Callable, num_scores: int):
+    module_wise: bool
+
+    module_wise_scores: dict[str, torch.Tensor]
+
+    module_wise_ssq: dict[str, torch.Tensor]
+
+    writer: ScoreWriter
+
+    def __init__(
+        self,
+        callback: Callable,
+        num_scores: int,
+        writer: ScoreWriter,
+        module_wise: bool = False,
+    ):
         self.callback = callback
         self.num_scores = num_scores
+        self.writer = writer
+        self.module_wise = module_wise
+        self.module_wise_scores = {}
+        self.module_wise_ssq = {}
 
-    def __call__(self, mod_grads: dict[str, torch.Tensor], **kwargs):
-        return self.callback(mod_grads, **kwargs)
+    def __call__(
+        self, indices: list[int], mod_grads: dict[str, torch.Tensor], **kwargs
+    ):
+        if self.module_wise:
+            name = kwargs["name"]
+            scores, ssq = self.callback(mod_grads, name=name)
 
+            self.module_wise_scores[name] = scores
+            if ssq is not None:
+                self.module_wise_ssq[name] = ssq
+        else:
+            scores = self.callback(mod_grads, **kwargs)
+            self.writer(indices, scores)
 
-def preprocess_grads(
-    grad_ds: Dataset,
-    grad_column_names: list[str],
-    unit_normalize: bool,
-    batch_size: int,
-    device: torch.device,
-    dtype: torch.dtype,
-    accumulate_grads: Literal["mean", "sum", "none"] = "none",
-    normalize_accumulated_grad: bool = False,
-) -> dict[str, torch.Tensor]:
-    """Preprocess the gradients in the dataset. Returns a dictionary
-    of preprocessed gradients with shape [1, grad_dim]. Preprocessing
-    includes some combination of unit normalization, accumulation,
-    accumulated gradient normalization, and dtype conversion."""
+    def finalize_module_wise(self, indices: list[int]):
+        if not self.module_wise:
+            return
 
-    # Short-circuit if possible
-    if accumulate_grads == "none" and not unit_normalize:
-        return {
-            column_name: grad_ds[:][column_name].to(device=device, dtype=dtype)
-            for column_name in grad_column_names
-        }
+        # Accumulate an inner product for each index
+        # [num_items, num_scores, num_modules] -> [num_items, num_scores]
+        scores = torch.stack(
+            [scores for scores in self.module_wise_scores.values()], dim=-1
+        ).sum(dim=-1)
 
-    # Get sum and sum of squares of the gradients
-    acc = {
-        column_name: torch.zeros_like(
-            grad_ds[0][column_name], device=device, dtype=torch.float32
-        )
-        for column_name in grad_column_names
-    }
-    ss_acc = torch.tensor(0.0, device=device, dtype=torch.float32)
-    if not unit_normalize:
-        ss_acc.fill_(1.0)
+        # Normalize with the sum of squares if present
+        if self.module_wise_ssq:
+            # [num_items, num_scores, num_modules] -> [num_items, num_scores]
+            sum_of_squares = torch.stack(
+                [sum_of_squares for sum_of_squares in self.module_wise_ssq.values()],
+                dim=-1,
+            ).sum(dim=-1)
+            assert scores.shape[0] == sum_of_squares.shape[0]
+            scores *= sum_of_squares.rsqrt()
 
-    def sum_(cols):
-        nonlocal ss_acc
-
-        for column_name in grad_column_names:
-            x = cols[column_name].to(device=device, dtype=torch.float32)
-            acc[column_name].add_(x.sum(0))
-
-            if unit_normalize:
-                # To normalize the mean gradient we can divide by the sum of
-                # squares of every gradient element in the dataset
-                ss_acc += x.pow(2).sum()
-
-    grad_ds.map(
-        sum_,
-        batched=True,
-        batch_size=batch_size,
-    )
-
-    ss_acc = ss_acc.sqrt()
-    assert ss_acc > 0, "Sum of squares of entire dataset is zero"
-
-    # Process the gradient dataset
-    if accumulate_grads == "mean":
-        grads = {
-            column_name: (acc[column_name] / ss_acc / len(grad_ds))
-            .unsqueeze(0)
-            .to(dtype)
-            for column_name in grad_column_names
-        }
-    elif accumulate_grads == "sum":
-        grads = {
-            column_name: (acc[column_name] / ss_acc).unsqueeze(0).to(dtype)
-            for column_name in grad_column_names
-        }
-    elif accumulate_grads == "none":
-        grads = {
-            column_name: grad_ds[:][column_name].to(device=device, dtype=dtype) / ss_acc
-            for column_name in grad_column_names
-        }
-    else:
-        raise ValueError(f"Invalid accumulate_grads: {accumulate_grads}")
-
-    # Normalize the accumulated gradient
-    if normalize_accumulated_grad:
-        grad_norm = torch.cat(
-            [grads[column_name].flatten() for column_name in grad_column_names], dim=0
-        ).norm()
-        for column_name in grad_column_names:
-            grads[column_name] /= grad_norm
-
-    return grads
+        self.writer(indices, scores)
 
 
-def build_scorer(
-    query_ds: Dataset,
+def build_scorer_callback(
+    query_grads: dict[str, torch.Tensor],
     query_cfg: QueryConfig,
     device: torch.device,
     dtype: torch.dtype,
     *,
-    accumulate_grads: Literal["mean", "sum", "none"] = "none",
     module_wise: bool = False,
     nearest: bool = False,
-    normalize_accumulated_grad: bool = False,
-) -> Scorer:
+) -> Callable:
     """Unified scorer builder for all scorer types."""
-
-    query_grads = preprocess_grads(
-        query_ds,
-        query_cfg.modules,
-        query_cfg.unit_normalize,
-        query_cfg.batch_size,
-        device,
-        dtype,
-        accumulate_grads=accumulate_grads,
-        normalize_accumulated_grad=normalize_accumulated_grad,
-    )
-
     if not module_wise:
         query_tensor = torch.cat(
             [query_grads[m].to(device=device, dtype=dtype) for m in query_cfg.modules],
@@ -145,9 +95,11 @@ def build_scorer(
             module_scores = mod_grads[name] @ query_grads[name].T
 
             # Save sum of squares for post-hoc normalization
-            # TODO only compute sum of squares if normalization requested
-            mod_grads[name].pow_(2)
-            ssq = mod_grads[name].sum(dim=1)
+            if query_cfg.unit_normalize:
+                mod_grads[name].pow_(2)
+                ssq = mod_grads[name].sum(dim=1)
+            else:
+                ssq = None
 
             return module_scores, ssq
 
@@ -161,110 +113,59 @@ def build_scorer(
 
         return grads @ query_tensor.T
 
-    num_scores = len(query_grads[query_cfg.modules[0]])
-
-    return Scorer(callback, num_scores)
-
-
-def get_query_ds(query_cfg: QueryConfig):
-    """
-    Load and optionally precondition the query dataset. Preconditioners
-    may be mixed as described in https://arxiv.org/html/2410.17413v1#S3.
-    """
-    # Collect the query gradients if they don't exist
-    query_path = Path(query_cfg.query_path)
-    if not query_path.exists():
-        raise FileNotFoundError(
-            f"Query dataset not found at {query_cfg.query_path}. "
-            "Please build a query dataset index first."
-        )
-
-    # Load the query dataset
-    with open(query_path / "info.json", "r") as f:
-        target_modules = json.load(f)["dtype"]["names"]
-
-    if not query_cfg.modules:
-        query_cfg.modules = target_modules
-
-    query_ds = load_gradient_dataset(
-        Path(query_cfg.query_path), concatenate_gradients=False
-    )
-    query_ds = query_ds.with_format("torch", columns=target_modules)
-
-    use_q = query_cfg.query_preconditioner_path is not None
-    use_i = query_cfg.index_preconditioner_path is not None
-
-    if use_q or use_i:
-        q, i = {}, {}
-        if use_q:
-            assert query_cfg.query_preconditioner_path is not None
-            q = GradientProcessor.load(
-                Path(query_cfg.query_preconditioner_path),
-                map_location="cuda",
-            ).preconditioners
-        if use_i:
-            assert query_cfg.index_preconditioner_path is not None
-            i = GradientProcessor.load(
-                Path(query_cfg.index_preconditioner_path), map_location="cuda"
-            ).preconditioners
-
-        mixed_preconditioner = (
-            {
-                k: q[k] * query_cfg.mixing_coefficient
-                + i[k] * (1 - query_cfg.mixing_coefficient)
-                for k in q
-            }
-            if (q and i)
-            else (q or i)
-        )
-        mixed_preconditioner = {k: v.cuda() for k, v in mixed_preconditioner.items()}
-
-        def precondition(batch):
-            for name in target_modules:
-                batch[name] = (batch[name].cuda() @ mixed_preconditioner[name]).cpu()
-
-            return batch
-
-        query_ds = query_ds.map(
-            precondition, batched=True, batch_size=query_cfg.batch_size
-        )
-
-    return query_ds.with_format("torch", columns=query_cfg.modules)
+    return callback
 
 
 def get_scorer(
-    query_cfg: QueryConfig, module_wise: bool, device: torch.device, dtype: torch.dtype
-) -> Scorer:
-    query_ds = get_query_ds(query_cfg)
+    query_grads: dict[str, torch.Tensor],
+    query_cfg: QueryConfig,
+    writer: ScoreWriter,
+    module_wise: bool,
+    device: torch.device,
+    dtype: torch.dtype,
+):
+    num_scores = len(query_grads[query_cfg.modules[0]])
 
     if query_cfg.score == "mean":
-        return build_scorer(
-            query_ds,
-            query_cfg,
-            device,
-            dtype,
+        return Scorer(
+            build_scorer_callback(
+                query_grads,
+                query_cfg,
+                device,
+                dtype,
+                module_wise=module_wise,
+            ),
+            num_scores,
+            writer,
             module_wise=module_wise,
-            accumulate_grads="mean",
         )
     elif query_cfg.score == "nearest":
         assert not module_wise, "Module-wise scoring not supported for nearest query"
-        return build_scorer(
-            query_ds,
-            query_cfg,
-            device,
-            dtype,
-            accumulate_grads="none",
-            module_wise=False,
-            nearest=True,
+        return Scorer(
+            build_scorer_callback(
+                query_grads,
+                query_cfg,
+                device,
+                dtype,
+                module_wise=False,
+                nearest=True,
+            ),
+            num_scores,
+            writer,
+            module_wise=module_wise,
         )
     elif query_cfg.score == "individual":
-        return build_scorer(
-            query_ds,
-            query_cfg,
-            device,
-            dtype,
+        return Scorer(
+            build_scorer_callback(
+                query_grads,
+                query_cfg,
+                device,
+                dtype,
+                module_wise=module_wise,
+            ),
+            num_scores,
+            writer,
             module_wise=module_wise,
-            accumulate_grads="none",
         )
     else:
         raise ValueError(f"Invalid scoring method: {query_cfg.score}")
