+++ tests/test_query.py
@@ -2,7 +2,6 @@ from pathlib import Path
 
 import pytest
 import torch
-from datasets import Dataset
 
 from bergson import (
     GradientCollector,
@@ -11,7 +10,7 @@ from bergson import (
     collect_gradients,
 )
 from bergson.data import QueryConfig
-from bergson.scorer import build_scorer
+from bergson.scorer import get_scorer
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
@@ -19,32 +18,27 @@ def test_query(tmp_path: Path, model, dataset):
     processor = GradientProcessor(projection_dim=16)
     shapes = GradientCollector(model.base_model, lambda x: x, processor).shapes()
 
-    query_gradient_ds = Dataset.from_list(
-        [
-            {module: torch.randn(shape.numel()) for module, shape in shapes.items()}
-            for _ in range(2)
-        ]
-    ).with_format("torch", columns=list(shapes.keys()))
+    query_grads = {
+        module: torch.randn(1, shape.numel()) for module, shape in shapes.items()
+    }
 
-    scorer = build_scorer(
-        query_gradient_ds,
+    score_writer = MemmapScoreWriter(
+        tmp_path,
+        len(dataset),
+        1,
+        rank=0,
+    )
+
+    scorer = get_scorer(
+        query_grads,
         QueryConfig(
             query_path=str(tmp_path / "query_gradient_ds"),
             modules=list(shapes.keys()),
             score="mean",
         ),
+        writer=score_writer,
         device=torch.device("cpu"),
         dtype=torch.float32,
-        module_wise=False,
-        accumulate_grads="mean",
-        normalize_accumulated_grad=True,
-    )
-    score_writer = MemmapScoreWriter(
-        scorer,
-        len(dataset),
-        tmp_path,
-        rank=0,
-        modules=list(shapes.keys()),
         module_wise=False,
     )
 
@@ -53,8 +47,7 @@ def test_query(tmp_path: Path, model, dataset):
         data=dataset,
         processor=processor,
         path=tmp_path,
-        score_writer=score_writer,
-        module_wise=False,
+        scorer=scorer,
     )
 
     assert any(tmp_path.iterdir()), "Expected artifacts in the temp run_path"
@@ -66,33 +59,28 @@ def test_module_wise_query(tmp_path: Path, model, dataset):
     processor = GradientProcessor(projection_dim=16)
     shapes = GradientCollector(model.base_model, lambda x: x, processor).shapes()
 
-    query_gradient_ds = Dataset.from_list(
-        [
-            {module: torch.randn(shape.numel()) for module, shape in shapes.items()}
-            for _ in range(2)
-        ]
-    ).with_format("torch", columns=list(shapes.keys()))
+    query_grads = {
+        module: torch.randn(1, shape.numel()) for module, shape in shapes.items()
+    }
 
-    scorer = build_scorer(
-        query_gradient_ds,
+    score_writer = MemmapScoreWriter(
+        tmp_path,
+        len(dataset),
+        1,
+        rank=0,
+    )
+
+    scorer = get_scorer(
+        query_grads,
         QueryConfig(
             query_path=str(tmp_path / "query_gradient_ds"),
             modules=list(shapes.keys()),
             score="mean",
         ),
+        writer=score_writer,
+        module_wise=True,
         device=torch.device("cpu"),
         dtype=torch.float32,
-        module_wise=True,
-        accumulate_grads="mean",
-        normalize_accumulated_grad=True,
-    )
-    score_writer = MemmapScoreWriter(
-        scorer,
-        len(dataset),
-        tmp_path,
-        rank=0,
-        modules=list(shapes.keys()),
-        module_wise=True,
     )
 
     collect_gradients(
@@ -100,7 +88,7 @@ def test_module_wise_query(tmp_path: Path, model, dataset):
         data=dataset,
         processor=processor,
         path=tmp_path,
-        score_writer=score_writer,
+        scorer=scorer,
         module_wise=True,
     )
